Fix ScheduleManager: Allow ops to be processed up to incomplete batch (#8206)
After PR #8137, it's now easier to see what's wrong with ScheduleManager.
You can see impact of the wrong doing described in detail in issue #7312

The core of the problem is that ScheduleManager will pause processing on incomplete batch coming in. Even if we have plenty of ops before that incomplete batch that could be processed just fine.
What likely happens (that makes it really bad in production) is that we could get multiple chunks of ops from storage, but each chunk might end up splitting some batch. Because we always have incomplete batch, we keep inbound queue in paused state until this problem resolves by fetching chunk of ops that does not break some batch.

This change also ensures that we pause inbound queue only when we have partial batch and no other ops.
It will not pause queue if inbound queue is empty, thus reducing number of paused - resumed transitions if we are getting non-batched ops.